Finish refactoring (idiot)
