+++ src/main/java/net/lukkit/lukkit/sandbox/api/http/method/POST.java
@@ -1,5 +1,5 @@
-package net.lukkit.lukkit.sandbox.globals.http.method;
+package net.lukkit.lukkit.sandbox.api.http.method;
 
-import net.lukkit.lukkit.sandbox.globals.http.BaseRequest;
+import net.lukkit.lukkit.sandbox.api.http.BaseRequest;
 
 public class POST extends BaseRequest {}
